DolphinProcess: Determine the presence of MEM2 automatically
This also remove any possibility to change whether to consider MEM2 or not.  This is possible because Dolphin maps MEM2 at a fixed offset from MEM1 in the host's RAM and it is only mapped when a Wii game is booted.   Since this makes everything more user firendly, the UI now labels MEM2 as "extended Wii-only memory/RAM" so it is clear that this should only be present for wii games.  The label telling the status of MEM2 is now just saying if it is present or absent when we are hooked to Dolphin.  No user intervention should be necessary after this commit to determine the status of MEM2.