Add Visual Studio Solution project files
This has the Qt build stems set for the files that need to be generated using Qt's MOC.
The project requires the variable $(QTDIR) to be set, which is automatically set when the
Qt VS Tools plugin is installed, and a path has been specified. If not, it can be added
manually to the .user file.