Use absolute paths
Relative paths were annoying to work with when Visual Studio doesn't gives an
application a working directory which is not the same as the output location.
This commit makes sure paths are absolute (even with a ..\ in there), so that
it doesn't matter where the output will be. It also adds a trailing backslash
behind paths to directories, and simplifies the project file somewhat.