Make project file smarter by using wildcards and properties
Moc now generates the source files inside a folder with the configuration name
(GeneratedFiles/Debug and GeneratedFiles/Release). The files that need to be
generated are alreasy present in the solution file, even when they don't exist
yet, but since the tool runs before the files get compiled it's not an issue.
This commit also gets rid of lots of duplicated settings (mainly for the moc
commands), and passes all preprocessor definitions to it.
This also adds quotes around the path, so that it may contain spaces.