MacDolphinProcess: Implement reading and writing
Performance is very poor with CPU usage getting up to 100% when idle. 90% of this time is spent in updating the cache.

The method of finding the correct memory region isn't perfect, but I think it will work.

Because it modifies other tasks, the executable must be signed with entitlements.xml