Reinstate **View > Scanner** top-bar menu action.
This checkable menu action was removed in dd7a5adab2ba9405d, as it was
deemed redundant after the main vertical splitter was added in
76e3b2eb6e77d9e42. However, it turned out that this convenience action
was quite popular, and users have requested it back.

An effort has been made to preserve the splitter's position when the
scanner's visibility is toggled.

Bonus: The action has been assigned the `F3` keyboard shortcut.

Fixes #102.