Support for multiple value edits.
When multiple watch entries are selected, and an edit is applied to one
of the values, the new value is set in all addresses (provided the
watch entry has the same type).

To edit a value without losing the multi selection, `F2` can be pressed
while the focus is set on a value cell.

Fixes #48.