Remove remnants of the deactivated sorting mechanism.
This sorting mechanism was deactivated in 65bb529fbd6f52ac547e, but the
actual functionality was still in the code base.

Note that sorting was disabled due to its implementation being
destructive: the actual watch nodes were reordered; not just the visual
representation of the items in the Qt tree view. If the functionality
is reinstated in the future, a `QSortFilterProxyModel` should be used.