Add fallback for external Qt in CMake file on Windows
On a clean build using CMake on Windows without Qt6 installed, CMakeLists did not use the external library for Qt 6.5.3 provided with the source. On windows, this work adds a quiet check for Qt6, then if it is not available it adds the external library to the prefix path list and checks again for Qt6.