Fixed bug on Exact Search for Assembly
Since `blr` can be represented as 0x4FFF0020 and 0x4E800020, the exact search on the memory scanner needs to account for this. So the memory is turned into an assembly instruction to see if it matches instead of comparing the raw memory. We check the raw memory if the opcode is 0 or 1 for performance reasons as this would take a long time otherwise.

Other search types (like greater than) aren't affected by this and I don't think there is a way to solve that