Adjust fmt in Cmake file
So before it worked, but linux would make a folder in `./Source/Externals/fmt` which shouldn't exist. This new method has linux and mac just install it as a dependency, and windows will compile it itself then link to it.

`submodules: recursive` was something I added for the previous way I got the code to compile